Buttons are re-usable and can change chart component depending on props
