+++ public/data/Spending Distribution/2020.csv
@@ -1,6 +1,6 @@
 name,value
 Food,11.90
-Delivery Fees,34.90
+Housing,34.90
 Transportation,16.00
 Healthcare,8.40
 All other expenditures,28.80
