+++ public/data/Spending Distribution/2021.csv
@@ -1,0 +1,6 @@
+name,value
+Food,12.40
+Housing,33.80
+Transportation,16.40
+Healthcare,8.10
+All other expenditures,29.30
