+++ public/data/Spending Distribution/2022.csv
@@ -1,0 +1,6 @@
+name,value
+Food,12.80
+Housing,33.30
+Transportation,16.80
+Healthcare,8.00
+All other expenditures,29.10
