+++ public/data/Spending Distribution/2023.csv
@@ -1,0 +1,6 @@
+name,value
+Food,12.90
+Housing,32.90
+Transportation,17.00
+Healthcare,8.00
+All other expenditures,29.20
